+++ workflow/lib/merge/stitch_well.py
@@ -782,86 +782,7 @@ class AlignedTiffEdge:
         return offset_tiff(tile_a, tile_b, self.relation, overlap=overlap)
 
 
-def estimate_stitch_aligned_tiff(
-    metadata_df: pd.DataFrame,
-    well: str,
-    data_type: str = "phenotype",
-    flipud: bool = False,
-    fliplr: bool = False,
-    rot90: int = 0,
-    channel: int = 0,
-    tile_size: Optional[Tuple[int, int]] = None,
-) -> Dict[str, Dict]:
-    """Complete stitching estimation preserving actual well geometry."""
-    # Use the new geometry-preserving function
-    tile_positions = metadata_to_grid_positions(metadata_df, well)
 
-    if len(tile_positions) == 0:
-        print(f"No tiles found for well {well}")
-        return {"total_translation": {}, "confidence": {well: {}}}
-
-    # Auto-detect tile size if not provided
-    if tile_size is None:
-        if data_type == "phenotype":
-            tile_size = (2400, 2400)  # Phenotype tiles
-        else:  # SBS
-            tile_size = (1200, 1200)  # SBS tiles
-        print(f"Auto-detected tile size for {data_type}: {tile_size}")
-
-    # Get actual stage coordinates for better connectivity
-    well_metadata = metadata_df[metadata_df["well"] == well].copy()
-    coords = well_metadata[["x_pos", "y_pos"]].values
-    tile_ids = well_metadata["tile"].values
-
-    # Use the improved connectivity function
-    edges = connectivity_from_actual_positions_optimized(
-        tile_positions, coords, tile_ids, data_type
-    )
-    if len(edges) == 0:
-        print("Warning: No edges created - tiles may be too far apart")
-        # Fallback to simple sequential connectivity
-        tile_list = list(tile_positions.keys())
-        edges = {}
-        for i in range(len(tile_list) - 1):
-            pos_a = tile_positions[tile_list[i]]
-            pos_b = tile_positions[tile_list[i + 1]]
-            edges[f"{i}"] = [pos_a, pos_b]
-        print(f"Created {len(edges)} fallback sequential edges")
-
-    print(f"Created {len(edges)} connectivity edges for well geometry")
-
-    tile_cache = AlignedTiffTileCache(
-        metadata_df, well, data_type, flipud, fliplr, rot90, channel
-    )
-
-    # Process edges
-    edge_list = []
-    confidence_dict = {}
-    pos_to_tile = {pos: tile_id for tile_id, pos in tile_positions.items()}
-
-    print(f"Processing {len(edges)} edges for {data_type} stitching...")
-    for key, (pos_a, pos_b) in tqdm(
-        edges.items(), desc=f"Computing pairwise shifts for {data_type}"
-    ):
-        tile_a_id = pos_to_tile[pos_a]
-        tile_b_id = pos_to_tile[pos_b]
-
-        try:
-            edge = AlignedTiffEdge(tile_a_id, tile_b_id, tile_cache, tile_positions)
-            edge_list.append(edge)
-            confidence_dict[key] = [
-                list(pos_a),
-                list(pos_b),
-                float(edge.model.confidence),
-            ]
-        except Exception as e:
-            print(f"Failed to process edge {tile_a_id}-{tile_b_id}: {e}")
-            continue
-
-    # Calculate optimal positions (this should now preserve well geometry)
-    opt_shift_dict = optimal_positions_tiff(edge_list, tile_positions, well, tile_size)
-
-    return {"total_translation": opt_shift_dict, "confidence": {well: confidence_dict}}
 
 
 def optimal_positions_tiff(
@@ -1617,297 +1538,7 @@ def verify_cell_id_preservation(
         traceback.print_exc()
 
 
-def estimate_stitch_sbs_coordinate_based(
-    metadata_df: pd.DataFrame,
-    well: str,
-    flipud: bool = False,
-    fliplr: bool = False,
-    rot90: int = 0,
-    channel: int = 0,
-) -> Dict[str, Dict]:
-    """Coordinate-based stitching for SBS data with correct scaling."""
-    well_metadata = metadata_df[metadata_df["well"] == well].copy()
-
-    if len(well_metadata) == 0:
-        print(f"No SBS tiles found for well {well}")
-        return {"total_translation": {}, "confidence": {well: {}}}
-
-    coords = well_metadata[["x_pos", "y_pos"]].values
-    tile_ids = well_metadata["tile"].values
-    tile_size = (1200, 1200)  # SBS tile size in pixels
-
-    print(f"Creating coordinate-based SBS stitch config for {len(tile_ids)} tiles")
-
-    # Use proven spacing detection
-    from scipy.spatial.distance import pdist
-
-    distances = pdist(coords)
-    actual_spacing = np.percentile(distances[distances > 0], 10)  # 10th percentile
-
-    print(f"SBS detected spacing: {actual_spacing:.1f} μm")
-    print(f"SBS tile size: {tile_size} pixels")
-
-    # FIXED: Use pixel size from metadata
-    if (
-        "pixel_size_x" in well_metadata.columns
-        and "pixel_size_y" in well_metadata.columns
-    ):
-        # Use pixel size from metadata (in μm per pixel)
-        pixel_size_um = well_metadata["pixel_size_x"].iloc[0]  # μm per pixel
-        pixels_per_micron = 1.0 / pixel_size_um  # pixels per μm
-        print(f"SBS pixel size from metadata: {pixel_size_um:.6f} μm/pixel")
-        print(f"SBS pixels per micron: {pixels_per_micron:.4f}")
-
-        # Verify pixel_size_y matches pixel_size_x
-        pixel_size_y = well_metadata["pixel_size_y"].iloc[0]
-        if abs(pixel_size_um - pixel_size_y) > 1e-6:
-            print(
-                f"⚠️  Warning: pixel_size_x ({pixel_size_um:.6f}) != pixel_size_y ({pixel_size_y:.6f})"
-            )
-    else:
-        print(
-            "⚠️  pixel_size_x/y not found in metadata, falling back to calculated values"
-        )
-        # Fallback: SBS specs: 1560 μm field of view, 1200 pixels -> 0.7692 pixels/μm
-        sbs_field_of_view_um = 1560.0  # μm
-        pixels_per_micron = tile_size[0] / sbs_field_of_view_um
-        print(f"SBS fallback - field of view: {sbs_field_of_view_um} μm")
-        print(f"SBS fallback - pixels per micron: {pixels_per_micron:.4f}")
-
-    x_min, y_min = coords.min(axis=0)
-
-    total_translation = {}
-    confidence = {}
-
-    for i, tile_id in enumerate(tile_ids):
-        x_pos, y_pos = coords[i]
-
-        # Convert directly to pixel coordinates using correct scale
-        pixel_x = int((x_pos - x_min) * pixels_per_micron)
-        pixel_y = int((y_pos - y_min) * pixels_per_micron)
-
-        total_translation[f"{well}/{tile_id}"] = [pixel_y, pixel_x]
-
-        # High confidence since using direct coordinates
-        confidence[f"coord_{i}"] = [[pixel_y, pixel_x], [pixel_y, pixel_x], 0.9]
-
-    print(f"Generated {len(total_translation)} SBS coordinate-based positions")
-
-    # Verify output size and spacing
-    y_shifts = [shift[0] for shift in total_translation.values()]
-    x_shifts = [shift[1] for shift in total_translation.values()]
-
-    if len(y_shifts) > 1:
-        pixel_spacings = []
-        for i in range(len(coords)):
-            for j in range(i + 1, len(coords)):
-                stage_dist = np.sqrt(
-                    (coords[i][0] - coords[j][0]) ** 2
-                    + (coords[i][1] - coords[j][1]) ** 2
-                )
-                pixel_dist = np.sqrt(
-                    (y_shifts[i] - y_shifts[j]) ** 2 + (x_shifts[i] - x_shifts[j]) ** 2
-                )
-                if stage_dist > 0:
-                    pixel_spacings.append(pixel_dist / stage_dist)
-
-        if pixel_spacings:
-            avg_pixel_spacing = np.mean(pixel_spacings)
-            print(
-                f"Verification - Average pixel spacing ratio: {avg_pixel_spacing:.4f} pixels/μm"
-            )
-
-            # Check for expected tile overlap
-            expected_tile_spacing_pixels = actual_spacing * pixels_per_micron
-            actual_avg_spacing = np.mean(
-                [
-                    np.sqrt(
-                        (y_shifts[i] - y_shifts[j]) ** 2
-                        + (x_shifts[i] - x_shifts[j]) ** 2
-                    )
-                    for i in range(len(y_shifts))
-                    for j in range(i + 1, len(y_shifts))
-                ]
-            )
-
-            if actual_avg_spacing > 0:
-                overlap_percent = (
-                    (tile_size[0] - actual_avg_spacing) / tile_size[0] * 100
-                )
-                print(f"SBS tile overlap: {overlap_percent:.1f}%")
-                if overlap_percent < 0:
-                    print("⚠️  Warning: Negative overlap detected - tiles may have gaps")
-                elif overlap_percent > 50:
-                    print(
-                        "⚠️  Warning: Very high overlap detected - may indicate scaling issues"
-                    )
-
-    final_size = (max(y_shifts) + tile_size[0], max(x_shifts) + tile_size[1])
-    memory_gb = final_size[0] * final_size[1] * 2 / 1e9
-
-    print(f"SBS final image size: {final_size}")
-    print(f"SBS memory estimate: {memory_gb:.1f} GB")
-
-    return {"total_translation": total_translation, "confidence": {well: confidence}}
 
 
-def estimate_stitch_phenotype_coordinate_based(
-    metadata_df: pd.DataFrame,
-    well: str,
-    flipud: bool = False,
-    fliplr: bool = False,
-    rot90: int = 0,
-    channel: int = 0,
-) -> Dict[str, Dict]:
-    """Coordinate-based stitching for phenotype data with correct scaling."""
-    well_metadata = metadata_df[metadata_df["well"] == well].copy()
-
-    if len(well_metadata) == 0:
-        print(f"No phenotype tiles found for well {well}")
-        return {"total_translation": {}, "confidence": {well: {}}}
-
-    coords = well_metadata[["x_pos", "y_pos"]].values
-    tile_ids = well_metadata["tile"].values
-    tile_size = (2400, 2400)  # Phenotype tile size in pixels
-
-    print(
-        f"Creating coordinate-based phenotype stitch config for {len(tile_ids)} tiles"
-    )
-
-    # Use proven spacing detection
-    from scipy.spatial.distance import pdist
-
-    distances = pdist(coords)
-    actual_spacing = np.percentile(distances[distances > 0], 10)  # 10th percentile
-
-    print(f"Phenotype detected spacing: {actual_spacing:.1f} μm")
-    print(f"Phenotype tile size: {tile_size} pixels")
-
-    # FIXED: Use pixel size from metadata
-    if (
-        "pixel_size_x" in well_metadata.columns
-        and "pixel_size_y" in well_metadata.columns
-    ):
-        # Use pixel size from metadata (in μm per pixel)
-        pixel_size_um = well_metadata["pixel_size_x"].iloc[0]  # μm per pixel
-        pixels_per_micron = 1.0 / pixel_size_um  # pixels per μm
-        print(f"Phenotype pixel size from metadata: {pixel_size_um:.6f} μm/pixel")
-        print(f"Phenotype pixels per micron: {pixels_per_micron:.4f}")
-
-        # Verify pixel_size_y matches pixel_size_x
-        pixel_size_y = well_metadata["pixel_size_y"].iloc[0]
-        if abs(pixel_size_um - pixel_size_y) > 1e-6:
-            print(
-                f"⚠️  Warning: pixel_size_x ({pixel_size_um:.6f}) != pixel_size_y ({pixel_size_y:.6f})"
-            )
-    else:
-        print(
-            "⚠️  pixel_size_x/y not found in metadata, falling back to calculated values"
-        )
-        # Fallback: Phenotype specs: 260 μm field of view, 2400 pixels -> 9.2308 pixels/μm
-        phenotype_field_of_view_um = 260.0  # μm
-        pixels_per_micron = tile_size[0] / phenotype_field_of_view_um
-        print(f"Phenotype fallback - field of view: {phenotype_field_of_view_um} μm")
-        print(f"Phenotype fallback - pixels per micron: {pixels_per_micron:.4f}")
-
-    x_min, y_min = coords.min(axis=0)
-
-    total_translation = {}
-    confidence = {}
-
-    for i, tile_id in enumerate(tile_ids):
-        x_pos, y_pos = coords[i]
-
-        # Convert directly to pixel coordinates using correct scale
-        pixel_x = int((x_pos - x_min) * pixels_per_micron)
-        pixel_y = int((y_pos - y_min) * pixels_per_micron)
-
-        total_translation[f"{well}/{tile_id}"] = [pixel_y, pixel_x]
-
-        # High confidence since using direct coordinates
-        confidence[f"coord_{i}"] = [[pixel_y, pixel_x], [pixel_y, pixel_x], 0.9]
-
-    print(f"Generated {len(total_translation)} phenotype coordinate-based positions")
-
-    # Verify output size and spacing
-    y_shifts = [shift[0] for shift in total_translation.values()]
-    x_shifts = [shift[1] for shift in total_translation.values()]
-
-    if len(y_shifts) > 1:
-        pixel_spacings = []
-        for i in range(len(coords)):
-            for j in range(i + 1, len(coords)):
-                stage_dist = np.sqrt(
-                    (coords[i][0] - coords[j][0]) ** 2
-                    + (coords[i][1] - coords[j][1]) ** 2
-                )
-                pixel_dist = np.sqrt(
-                    (y_shifts[i] - y_shifts[j]) ** 2 + (x_shifts[i] - x_shifts[j]) ** 2
-                )
-                if stage_dist > 0:
-                    pixel_spacings.append(pixel_dist / stage_dist)
-
-        if pixel_spacings:
-            avg_pixel_spacing = np.mean(pixel_spacings)
-            print(
-                f"Verification - Average pixel spacing ratio: {avg_pixel_spacing:.4f} pixels/μm"
-            )
-
-            # Check for expected tile overlap
-            expected_tile_spacing_pixels = actual_spacing * pixels_per_micron
-            actual_avg_spacing = np.mean(
-                [
-                    np.sqrt(
-                        (y_shifts[i] - y_shifts[j]) ** 2
-                        + (x_shifts[i] - x_shifts[j]) ** 2
-                    )
-                    for i in range(len(y_shifts))
-                    for j in range(i + 1, len(y_shifts))
-                ]
-            )
-
-            if actual_avg_spacing > 0:
-                overlap_percent = (
-                    (tile_size[0] - actual_avg_spacing) / tile_size[0] * 100
-                )
-                print(f"Phenotype tile overlap: {overlap_percent:.1f}%")
-                if overlap_percent < 0:
-                    print("⚠️  Warning: Negative overlap detected - tiles may have gaps")
-                elif overlap_percent > 50:
-                    print(
-                        "⚠️  Warning: Very high overlap detected - may indicate scaling issues"
-                    )
-
-    final_size = (max(y_shifts) + tile_size[0], max(x_shifts) + tile_size[1])
-    memory_gb = final_size[0] * final_size[1] * 2 / 1e9
-
-    print(f"Phenotype final image size: {final_size}")
-    print(f"Phenotype memory estimate: {memory_gb:.1f} GB")
-
-    return {"total_translation": total_translation, "confidence": {well: confidence}}
 
 
-def estimate_stitch_data_type_specific(
-    metadata_df: pd.DataFrame,
-    well: str,
-    data_type: str = "phenotype",
-    flipud: bool = False,
-    fliplr: bool = False,
-    rot90: int = 0,
-    channel: int = 0,
-    tile_size: Optional[Tuple[int, int]] = None,
-) -> Dict[str, Dict]:
-    """Route to data-type-specific optimized functions with corrected coordinate scaling."""
-    if data_type == "phenotype":
-        return estimate_stitch_phenotype_coordinate_based(
-            metadata_df, well, flipud, fliplr, rot90, channel
-        )
-    elif data_type == "sbs":
-        return estimate_stitch_sbs_coordinate_based(
-            metadata_df, well, flipud, fliplr, rot90, channel
-        )
-    else:
-        # Fallback to original approach
-        return estimate_stitch_aligned_tiff(
-            metadata_df, well, data_type, flipud, fliplr, rot90, channel, tile_size
-        )
